+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ## geouy v0.2.6 (in progress)
 
-* Add Municipios 2010 and 2015 to metadata 
+* Add Municipios 2010 and 2015, and Territorial Offices of MIDES to metadata 
 * Update geocode_ide_uy() because server of IDEuy change
 * Add a reverse_ide_uy() function for reverse geocoding 
 * Change geoservice of "Departamentos"